Remove setTimeout as possible cause for bugs
There is a bug now that fix cards when they are not shown, apparently
when the user clicks a pair of cards right after clicking another pair
containing one of the currently clicked cards.

The bug could be being caused by improper use of setTimeout, so I
decided to remove it while I find out why was it causing the bug
(if so) or, alternatively, find another way of accomplishing what was
being done with it.